requirements.txt
alembic==1.14.0
blinker==1.8.2
Bootstrap-Flask==2.4.1
click==8.1.7
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
pillow==11.0.0
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.2
WTForms==3.2.1
